+++ src/org/magiclib/subsystems/drones/HoveringFormation.kt
@@ -4,6 +4,7 @@ import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.combat.ShipAPI
 import com.fs.starfarer.api.util.Misc
 import org.lazywizard.lazylib.MathUtils
+import org.magiclib.util.PIDController
 
 class HoveringFormation : DroneFormation() {
     override fun advance(ship: ShipAPI, drones: Map<ShipAPI, PIDController>, amount: Float) {
